Add 1.5x and 2.5x scaling modes

+++ src/EmuWindow.h
@@ -80,7 +80,7 @@ class EmuWindow : public EmuObject, public PalWindow
         void setDefaultWindowSize(int width, int height);
         void setCaption(std::string caption);
         void setFrameScale(FrameScale fs);
-        void setFixedYScale(int yScale);
+        void setFixedYScale(double yScale);
         void setFieldsMixing(FieldsMixing fm);
         void setWindowStyle(WindowStyle ws);
         void setSmoothing(SmoothingType smoothing);
@@ -124,7 +124,7 @@ class EmuWindow : public EmuObject, public PalWindow
         FieldsMixing m_fieldsMixing = FM_NONE;
         WindowStyle m_windowStyle = WS_AUTOSIZE;
         SmoothingType m_smoothing = ST_SHARP;
-        int m_scaleY;
+        double m_scaleY;
 
         int m_curImgWidth = 0;
         int m_curImgHeight = 0;
@@ -143,7 +143,7 @@ class EmuWindow : public EmuObject, public PalWindow
         unsigned m_curFrameNo = unsigned(-1);
         bool m_frameDrawn = false;
 
-        double calcBestAspectRatio(double srcAspectRatio, int scaleY);
+        double calcBestAspectRatio(double srcAspectRatio, double scaleY);
         /*bool m_overlay = false;
         double m_primaryFrameAspectRatio = 1;*/
 };
